refactors and changes to hooks
