+++ src/components/editor/raw/interactions.ts
@@ -6,6 +6,10 @@ import {
 	dispatchTagClick,
 	dispatchWikiLinkClick,
 } from "../markdown/editorEvents";
+import {
+	type FootnoteKind,
+	findFootnoteCounterpartOffset,
+} from "../markdown/footnote";
 import { parseWikiLink } from "../markdown/wikiLinkCodec";
 
 function toggleTask(view: EditorView, target: HTMLElement): boolean {
@@ -26,13 +30,32 @@ function toggleTask(view: EditorView, target: HTMLElement): boolean {
 	return true;
 }
 
+function scrollToRawFootnoteCounterpart(
+	view: EditorView,
+	id: string,
+	fromKind: FootnoteKind,
+): boolean {
+	const offset = findFootnoteCounterpartOffset(
+		view.state.doc.toString(),
+		id,
+		fromKind,
+	);
+	if (offset === null) return false;
+	view.dispatch({
+		selection: { anchor: offset },
+		effects: EditorView.scrollIntoView(offset, { y: "center" }),
+	});
+	view.focus();
+	return true;
+}
+
 export function createRawMarkdownEventHandlers(getRelPath: () => string) {
 	return {
 		mousedown: (event: MouseEvent) => {
 			const target = event.target as Element | null;
 			if (
 				target?.closest(
-					".cm-raw-task-checkbox, .cm-raw-wiki-link, .cm-raw-markdown-link, .cm-raw-tag",
+					".cm-raw-task-checkbox, .cm-raw-wiki-link, .cm-raw-markdown-link, .cm-raw-tag, .cm-raw-footnote",
 				)
 			) {
 				event.preventDefault();
@@ -43,6 +66,16 @@ export function createRawMarkdownEventHandlers(getRelPath: () => string) {
 			const target = event.target as Element | null;
 			const task = target?.closest<HTMLElement>(".cm-raw-task-checkbox");
 			if (task) return toggleTask(view, task);
+
+			const footnote = target?.closest<HTMLElement>(".cm-raw-footnote");
+			if (footnote?.dataset.footnoteId) {
+				const kind = footnote.dataset.footnoteKind === "def" ? "def" : "ref";
+				return scrollToRawFootnoteCounterpart(
+					view,
+					footnote.dataset.footnoteId,
+					kind,
+				);
+			}
 
 			const wikiLink = target?.closest<HTMLElement>(".cm-raw-wiki-link");
 			if (wikiLink) {
