+++ src/utils/dom.ts
@@ -1,0 +1,6 @@
+export function cssEscape(value: string): string {
+	if (typeof CSS !== "undefined" && typeof CSS.escape === "function") {
+		return CSS.escape(value);
+	}
+	return value.replace(/["\\]/g, "\\$&");
+}
